refactor: Name 클래스로 String 원시값 포장
