+++ src/main/java/blackjack/domain/Name.java
@@ -1,0 +1,39 @@
+package blackjack.domain;
+
+import java.util.Objects;
+
+public class Name {
+    private static final int MAX_NAME_LENGTH = 5;
+
+    private final String name;
+
+    public Name(String name) {
+        validate(name);
+        this.name = name;
+    }
+
+    private void validate(String name) {
+        validateBlankName(name);
+        validateNameLength(name);
+    }
+
+    private void validateBlankName(String name) {
+        if (Objects.isNull(name) || name.isBlank()) {
+            throw new IllegalArgumentException("[ERROR] 이름은 공백일 수 없습니다.");
+        }
+    }
+
+    private void validateNameLength(String name) {
+        if (name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("[ERROR] 이름은 " + MAX_NAME_LENGTH + "글자를 넘을 수 없습니다.");
+        }
+    }
+
+    public boolean isNameMatch(String name) {
+        return Objects.equals(this.name, name);
+    }
+
+    public String getName() {
+        return name;
+    }
+}
